% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvalPll_sptw.R
\name{crossvalPll_sptw}
\alias{crossvalPll_sptw}
\title{Cross-validation to estimate coefficients}
\usage{
crossvalPll_sptw(
  K = 3,
  y = NULL,
  X = NULL,
  Z = NULL,
  index = NULL,
  beta.init = NULL,
  gamma.init = NULL,
  alpha.init = NULL,
  id.list = NULL,
  l1_seq = NULL,
  l2_seq = NULL,
  lapMat = NULL,
  miter = NULL,
  tol = NULL,
  p = NULL,
  verbose = TRUE
)
}
\arguments{
\item{K}{number of folds}

\item{y}{observed response}

\item{X}{covariates for the mean model}

\item{Z}{covariates for the dispersion model}

\item{index}{incidence matrix}

\item{beta.init}{initial values for mean coefficients}

\item{gamma.init}{initial values for dispersion coefficients}

\item{alpha.init}{initial values for spatial coefficients}

\item{id.list}{index for splits or folds}

\item{l1_seq}{sequence for \eqn{\lambda_1} (use log-scale)}

\item{l2_seq}{sequence for \eqn{\lambda_2} (use log-scale)}

\item{lapMat}{penalty matrix---the graph Laplacian}

\item{miter}{maximum iterations}

\item{tol}{tolerance (eg, 1e-10)}

\item{p}{index parameter}

\item{index.y.0}{index for 0 entries}
}
\description{
Function to perform \eqn{K}-fold cross-validation to estimate model coefficients. This function leverages parallel computation to speed up the process.
}
