% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/del1afun.R
\name{del1afun}
\alias{del1afun}
\title{Internal function}
\usage{
del1afun(
  y = NULL,
  phi = NULL,
  p = NULL,
  xi = NULL,
  index.y.0 = NULL,
  epsilon = 1e-17
)
}
\arguments{
\item{y}{response}

\item{phi}{dispersion vector}

\item{p}{index parameter}

\item{xi}{index parameter}

\item{index.y.0}{index for 0 values in response}
}
\description{
Not for external use
}
\keyword{del1afun}
