% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fold_split.R
\name{fold_split}
\alias{fold_split}
\title{Function for creating K-folds for cross-validation}
\usage{
fold_split(K = 3, index = NULL)
}
\arguments{
\item{K}{number of folds}
}
\description{
Internal function. Do not use externally
}
