% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathMM_sptw.R
\name{pathMM_sptw}
\alias{pathMM_sptw}
\title{Warm start}
\usage{
pathMM_sptw(
  y = NULL,
  X = NULL,
  Z = NULL,
  index = NULL,
  index.y.0 = NULL,
  beta.init = NULL,
  gamma.init = NULL,
  alpha.init = NULL,
  lapMat = NULL,
  p = NULL,
  tol = 0.001,
  miter = 10000,
  l1_seq = NULL,
  l2_seq = NULL,
  verbose = TRUE
)
}
\arguments{
\item{y}{observed response}

\item{X}{covariates for the mean model}

\item{Z}{covariates for the dispersion model}

\item{index}{incidence matrix}

\item{index.y.0}{index for 0 entries}

\item{beta.init}{initial values for mean coefficients}

\item{gamma.init}{initial values for dispersion coefficients}

\item{alpha.init}{initial values for spatial coefficients}

\item{lapMat}{penalty matrix---the graph Laplacian}

\item{p}{index parameter}

\item{l1_seq}{sequence for \eqn{\lambda_1} (use log-scale)}

\item{l2_seq}{sequence for \eqn{\lambda_2} (use log-scale)}
}
\description{
Function for performing a grid search for searching optimal penalization, \eqn{\lambda_1I + \lambda_2(D-W)}, where \eqn{D} is the degree matrix and \eqn{W} is the adjacency matrix for the graph.
}
