% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_tweedie.R
\name{spatial_tweedie}
\alias{spatial_tweedie}
\title{Tweedie Compound Poisson-Gamma (CP-g) Double Generalized Linear Spatial Model}
\usage{
spatial_tweedie(
  y = NULL,
  X = NULL,
  Z = NULL,
  index = NULL,
  index.y.0 = NULL,
  beta.init = NULL,
  gamma.init = NULL,
  alpha.init = NULL,
  pen_mat = NULL,
  p = 1.5,
  tol = 0.001,
  miter = 10000,
  inf = FALSE,
  saddle.approx = TRUE,
  verbose = TRUE,
  p.update = TRUE
)
}
\arguments{
\item{y}{observed response}

\item{X}{covariates for the mean model}

\item{Z}{covariates for the dispersion model}

\item{index}{incidence matrix}

\item{index.y.0}{index for 0 entries}

\item{beta.init}{initial values for mean coefficients}

\item{gamma.init}{initial values for dispersion coefficients}

\item{alpha.init}{initial values for spatial coefficients}

\item{pen_mat}{penalty matrix---the graph Laplacian}
}
\description{

}
\keyword{spatial_tweedie}
